Make ButtonState an Enumerator
This seems like an ideal place to implement an enumerator. Sure, we may be at qualifiers right now, but I think it would be a more logical set-up for future builds. Note: I haven't checked other dependent files and changed to follow this standard, I was just scrolling through and saw this during some down time.